Generate a on-time performance report

Retrieves and combines data from Zonar and Versatrans and generates full OTP data. The full schedule and OTP data is returned, it will need to be summarized for public reporting.

Usage

otp_report(
  date = as.character(Sys.Date()),
  ampm = c("AM", "PM"),
  include_zonar = TRUE,
  cutoff_min = -25,
  cutoff_max = 60,
  rp_database = Sys.getenv("RP_DATABASE"),
  os_database = Sys.getenv("OS_DATABASE"),
  rp_odbc_name = Sys.getenv("RP_ODBC_NAME"),
  uncovered_url = Sys.getenv("UNCOVERED_URL"),
  mail_to = strsplit(Sys.getenv("MAIL_TO"), ",")[[1]],
  mail_from = Sys.getenv("MAIL_FROM"),
  TZ = "America/New_York",
  test = FALSE
)

Arguments

date

Character vector of length 1 giving the date in YYY-MM-DD format.

ampm

Character vector of length 1, either "AM", or "PM", use to specify morning or evening OTP report.

include_zonar

Logical vector of length 1. If TRUE zonar schedule data will be retrieved and used for OTP calculations.

cutoff_min

numeric vector of length one giving the minimum delay time in minutes that will be considered a valid trip arrival. Usually this will be a negative value, e.g., -25 or so.

cutoff_max

numeric vector of length one giving the maximum delay time in minutes that will be considered a vlid trip arrival. Usuall this will be positive and relatively large, e.g., 60 or so.

rp_database

Name of the RP database to connect to. Retrieved from RP_DATABASE environment variable by default but can be overridden here, see RVersatransRP package for details.

os_database

Name of the Onscreen database to connect to.

rp_odbc_name

Name of the Windows ODBC data source to use. Retrieved from RP_ODBC_NAME environment variable by default but can be overridden here, see RVersatransRP package for details.

uncovered_url

URL of a google sheet containing a list of uncovered trips. Must contain columns named "Date", "AM/PM", "Route Set", and "Bus".

mail_to

Character vector of email addresses to send notifications to.

mail_from

GMail address used to send notification emails.

TZ

The timezone used by the database server.

test

Logical vector of length one indicating whether to pull an abbreviated test data set. Even with this the time needed to test this function will be long.

Details

TODO

Examples

otp <- otp_report("2023-02-14", ampm = "AM", test = TRUE)
#> ! Using an auto-discovered, cached token.
#>   To suppress this message, modify your code or options to clearly consent to
#>   the use of a cached token.
#>   See gargle's "Non-interactive auth" vignette for more details:
#>   <https://gargle.r-lib.org/articles/non-interactive-auth.html>
#> ℹ The googlesheets4 package is using a cached token for
#>   izahn@bostonpublicschools.org.
#> ✔ Reading from SY22-23 Uncovered Tracking.
#> ✔ Range '2022-23 SY Uncovered Trip Data (starting 8/29)'.
#> NOTE: only pulling schedule data for the first five zones; omit 'test' parameter if you want everything
dplyr::glimpse(otp[0,])
#> Rows: 0
#> Columns: 34
#> $ RPVehicle           <chr> 
#> $ RPRoute             <chr> 
#> $ ExpectedTime        <dttm> 
#> $ Arrival             <dttm> 
#> $ delayTimeCombined   <int> 
#> $ ActualTime          <dttm> 
#> $ zonarActualTime     <dttm> 
#> $ DelayTime           <int> 
#> $ zonarDelayTime      <int> 
#> $ Direction           <chr> 
#> $ AnchorAbbrev        <chr> 
#> $ Date                <dttm> 
#> $ AMPM                <chr> 
#> $ OvernightGarage     <chr> 
#> $ Names               <chr> 
#> $ RouteName           <chr> 
#> $ load                <int> 
#> $ Days                <chr> 
#> $ RouteSetName        <chr> 
#> $ excludedShuttlesEtc <lgl> 
#> $ OriginalRPVehicle   <chr> 
#> $ Zone                <chr> 
#> $ zonarCategory       <chr> 
#> $ zonarDuration       <drtn>  secs
#> $ inZonar             <lgl> 
#> $ inOnscreen          <lgl> 
#> $ arrivalTimeSource   <chr> 
#> $ arrivalInWindow     <lgl> 
#> $ zoneInZonar         <lgl> 
#> $ OSVehicleOverride   <lgl> 
#> $ Uncovered           <lgl> 
#> $ cutoff_min          <dbl> 
#> $ cutoff_max          <dbl> 
#> $ TripCategory        <chr>